functions/requirements.txt
firebase-admin>=6.0.0
flask-cors>=4.0.0
flask>=2.3.0
pytesseract>=0.3.10
Pillow>=9.0.0
spacy>=3.5.0
requests>=2.28.0
unidecode>=1.3.0
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.6.0/de_core_news_sm-3.6.0-py3-none-any.whl
